// applets/jellyfin/api/model/baseItemKind.ts
/**
 * Generated by orval v6.11.1 🍺
 * Do not edit manually.
 * Jellyfin API
 * OpenAPI spec version: 10.8.8
 */

/**
 * The base item kind.
 */
export type BaseItemKind = (typeof BaseItemKind)[keyof typeof BaseItemKind];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BaseItemKind = {
  AggregateFolder: "AggregateFolder",
  Audio: "Audio",
  AudioBook: "AudioBook",
  BasePluginFolder: "BasePluginFolder",
  Book: "Book",
  BoxSet: "BoxSet",
  Channel: "Channel",
  ChannelFolderItem: "ChannelFolderItem",
  CollectionFolder: "CollectionFolder",
  Episode: "Episode",
  Folder: "Folder",
  Genre: "Genre",
  ManualPlaylistsFolder: "ManualPlaylistsFolder",
  Movie: "Movie",
  LiveTvChannel: "LiveTvChannel",
  LiveTvProgram: "LiveTvProgram",
  MusicAlbum: "MusicAlbum",
  MusicArtist: "MusicArtist",
  MusicGenre: "MusicGenre",
  MusicVideo: "MusicVideo",
  Person: "Person",
  Photo: "Photo",
  PhotoAlbum: "PhotoAlbum",
  Playlist: "Playlist",
  PlaylistsFolder: "PlaylistsFolder",
  Program: "Program",
  Recording: "Recording",
  Season: "Season",
  Series: "Series",
  Studio: "Studio",
  Trailer: "Trailer",
  TvChannel: "TvChannel",
  TvProgram: "TvProgram",
  UserRootFolder: "UserRootFolder",
  UserView: "UserView",
  Video: "Video",
  Year: "Year",
} as const;